backend/requirements.txt
psycopg2-binary
Flask==2.2.3
pystardog
pdfkit
flask-cors
rdflib
python-dotenv
Werkzeug==2.3.7
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
flask-testing